refactor(map): streamline map component state management by consolidating hooks and removing unused code
